Move accordion definition to a helper component
